pyproject.toml
[tool.poetry]
name = "django-one"
version = "0.1.0"
description = ""
authors = ["Carlos <cafesanu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.2"
djangorestframework = "^3.12.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.790"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
